Add files via upload
The food aggregator company has stored the data of the different orders made by the registered customers in their online portal. They want to analyze the data to get a fair idea about the demand of different restaurants which will help them in enhancing their customer experience. In this project I am the Data Scientist at Foodhub and the Data Science team has shared some of the key questions that need to be answered. I Performed the data analysis to find answers to these questions that will help the company to improve the business.